Rename Subscriber to EventSubscriber
