+++ docs/examples/extension/src/EventSubscriber/HistorySubscriber.php
@@ -11,13 +11,14 @@
  * that were distributed with this source code.
  */
 
-namespace Example\Subscriber;
+namespace Example\EventSubscriber;
 
 use Eventum\Event\SystemEvents;
+use Misc;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\EventDispatcher\GenericEvent;
 
-class UserSubscriber implements EventSubscriberInterface
+class HistorySubscriber implements EventSubscriberInterface
 {
     /**
      * {@inheritdoc}
@@ -25,24 +26,17 @@ class UserSubscriber implements EventSubscriberInterface
     public static function getSubscribedEvents(): array
     {
         return [
-            SystemEvents::USER_CREATE => 'userCreated',
-            SystemEvents::USER_UPDATE => 'userUpdated',
+            SystemEvents::HISTORY_ADD => 'historyAdded',
         ];
     }
 
     /**
      * @param GenericEvent $event
      */
-    public function userCreated(GenericEvent $event): void
+    public function historyAdded(GenericEvent $event): void
     {
-        error_log("user created: #{$event['id']}");
-    }
+        $his_summary = Misc::processTokens(ev_gettext($event['his_summary']), $event['his_context']);
 
-    /**
-     * @param GenericEvent $event
-     */
-    public function userUpdated(GenericEvent $event): void
-    {
-        error_log("user updated: #{$event['id']}");
+        error_log("HISTORY[issue {$event['his_iss_id']}]: {$event['his_id']}: $his_summary");
     }
 }
